test(upstream): refactors class exercise tests
- splits unit test suite in half for each part of exercise
- updates README instructions
