+++ exercises/object-oriented-programming/src/test/java/com/cbfacademy/cars/ShowroomTest.java
@@ -2,19 +2,14 @@ package com.cbfacademy.cars;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.HashMap;
 import java.util.List;
-import java.util.stream.Stream;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-@DisplayName(value = "Class exercises tests")
-public class ClassExercisesTest {
+@DisplayName(value = "Showroom tests")
+public class ShowroomTest {
 
     @Test
     @DisplayName("Showroom object has a valid getCars method")
@@ -25,50 +20,6 @@ public class ClassExercisesTest {
         for (HashMap<String, Object> expected : carValues) {
             assertTrue(hasMatch(cars, expected));
         }
-    }
-
-    @ParameterizedTest
-    @MethodSource("getCarArguments")
-    @DisplayName("Car object has valid getMake method")
-    public void testCarObjectHasValidGetMakeMethod(Car car, String make, String model, int year) {
-        assertEquals(make, car.getMake());
-    }
-
-    @ParameterizedTest
-    @MethodSource("getCarArguments")
-    @DisplayName("Car object has valid getModel method")
-    public void testCarObjectHasValidGetModelMethod(Car car, String make, String model, int year) {
-        assertEquals(model, car.getModel());
-    }
-
-    @ParameterizedTest
-    @MethodSource("getCarArguments")
-    @DisplayName("Car object has valid getYear method")
-    public void testCarObjectHasValidGetYearMethod(Car car, String make, String model, int year) {
-        assertEquals(year, car.getYear());
-    }
-
-    @ParameterizedTest
-    @MethodSource("getCarArguments")
-    @DisplayName("Car object has valid getDetails method")
-    public void testCarObjectHasValidGetDetailsMethod(Car car, String make, String model, int year) {
-        String details = car.getDetails();
-
-        assertTrue(details.indexOf(make) >= 0 && details.indexOf(model) >= 0
-                && details.indexOf(Integer.toString(year)) >= 0);
-    }
-
-    private static Stream<Arguments> getCarArguments() {
-        List<HashMap<String, Object>> cars = getCarValues();
-
-        return cars.stream()
-                .map(map -> {
-                    String make = (String) map.get("make");
-                    String model = (String) map.get("model");
-                    int year = (int) map.get("year");
-
-                    return Arguments.of(new Car(make, model, year), make, model, year);
-                });
     }
 
     private static List<HashMap<String, Object>> getCarValues() {
